fix rendering in kernel template
division will occasionaly leave more items in a column than
the 'num_per_col' variable, which leads to a duplicate array index
on the patch_status array.

As {% set %} doesn't modify things outside the current scope (ie, {% for %}),
set a 'counter' array and increment it in an if statement on every loop.

Change-Id: Iddc685b06fb1a13275bd4b4aee61882d83e2fd9b